Extract run
