+++ exo-go/src/run/service_restarter.go
@@ -1,4 +1,4 @@
-package application
+package run
 
 import (
 	"fmt"
